Switch to f64 to avoid artefacting

+++ src/camera.rs
@@ -7,7 +7,7 @@ use std::fs::File;
 use std::io::Write;
 
 pub struct Camera {
-    pub aspect_ratio: f32,
+    pub aspect_ratio: f64,
     pub img_width: u32,
     pub samples_per_pixel: u32,
     img_height: u32,
@@ -18,19 +18,19 @@ pub struct Camera {
 }
 
 impl Camera {
-    pub fn new(aspect_ratio: f32, img_width: u32, samples_per_pixel: u32) -> Camera {
-        let img_height: u32 = (img_width as f32 / aspect_ratio) as u32;
+    pub fn new(aspect_ratio: f64, img_width: u32, samples_per_pixel: u32) -> Camera {
+        let img_height: u32 = (img_width as f64 / aspect_ratio) as u32;
 
-        let focal_length: f32 = 1.0;
-        let viewport_height: f32 = 2.0;
-        let viewport_width: f32 = viewport_height * (img_width as f32 / img_height as f32);
+        let focal_length: f64 = 1.0;
+        let viewport_height: f64 = 2.0;
+        let viewport_width: f64 = viewport_height * (img_width as f64 / img_height as f64);
         let center = Vec3::new(0.0, 0.0, 0.0);
 
         let viewport_u = Vec3::new(viewport_width, 0.0, 0.0);
         let viewport_v = Vec3::new(0.0, -viewport_height, 0.0);
 
-        let pixel_delta_u = viewport_u / (img_width as f32);
-        let pixel_delta_v = viewport_v / (img_height as f32);
+        let pixel_delta_u = viewport_u / (img_width as f64);
+        let pixel_delta_v = viewport_v / (img_height as f64);
 
         let viewport_upper_left =
             center - Vec3::new(0.0, 0.0, focal_length) - viewport_u / 2.0 - viewport_v / 2.0;
@@ -50,14 +50,14 @@ impl Camera {
     }
 
     fn ray_color(ray: &Ray, world: &Vec<Box<dyn Hittable>>) -> Vec3 {
-        match world.hit(ray, Interval::new(0.0, f32::INFINITY)) {
+        match world.hit(ray, Interval::new(0.0, f64::INFINITY)) {
             Some(hit_rec) => {
                 let direction = Vec3::random_on_hemisphere(hit_rec.normal);
                 0.5 * Camera::ray_color(&Ray::new(hit_rec.point, direction), world)
             }
             None => {
                 let unit_direction = ray.direction.unit_vector();
-                let a: f32 = 0.5 * (unit_direction.y() + 1.0);
+                let a: f64 = 0.5 * (unit_direction.y() + 1.0);
 
                 (1.0 - a) * Vec3::new(1.0, 1.0, 1.0) + a * Vec3::new(0.5, 0.7, 1.0)
             }
@@ -65,7 +65,7 @@ impl Camera {
     }
 
     fn write_color(&self, mut file: &File, color: Vec3) {
-        let scale = 1.0 / self.samples_per_pixel as f32;
+        let scale = 1.0 / self.samples_per_pixel as f64;
 
         let r = color.x() * scale;
         let g = color.y() * scale;
@@ -92,7 +92,7 @@ impl Camera {
 
     fn get_ray(&self, i: u32, j: u32) -> Ray {
         let pixel_center =
-            self.pixel00_loc + (i as f32 * self.pixel_delta_u) + (j as f32 * self.pixel_delta_v);
+            self.pixel00_loc + (i as f64 * self.pixel_delta_u) + (j as f64 * self.pixel_delta_v);
         let pixel_sample = pixel_center + self.pixel_sample_square();
 
         let ray_origin = self.center;
