+++ src/hittable.rs
@@ -6,7 +6,7 @@ use crate::vec3::Vec3;
 pub struct HitRecord {
     pub point: Vec3,
     pub normal: Vec3,
-    pub t: f32,
+    pub t: f64,
     pub front_face: bool,
 }
